+++ Sources/CharcoalShared/Extensions/Rect+Extension.swift
@@ -1,11 +1,10 @@
 import Foundation
 
-
 public extension CGRect {
     // Calculate the intersection area of two rectangles
     func intersectionArea(_ rect: CGRect) -> CGFloat {
-        let rect = self.intersection(rect)
-        
+        let rect = intersection(rect)
+
         return rect.width * rect.height
     }
 }
