+++ Sources/CharcoalUIKit/Components/Overlay/CharcoalOverlayInteractionMode.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-enum CharcoalOverlayInteractionMode {
+public enum CharcoalOverlayInteractionMode {
     case passThrough
     case block
     case dimissOnTouch
